requirements.txt
aiohttp
hikari
hikari-lightbulb
python-dotenv
uvloop
